Added backup functionality
-- Updated README with description of what backup functionality does

-- Put backup folder in .gitignore (as well as *.pyc files)

-- Updated config.ini.template to have backup Retention field

-- Added backup.py to backup tasks.ini after every run
    -- If there is already say 3 files in /backup/ and the Retention field
        of config.ini is 3, it will delete the oldest backup and create a new backup
    -- backup files are saved in /backup/ with filename tasks.ini.backup.TIME
        where TIME is the number of seconds since the epoch
-- Called backup.createBackup() inside checkTasks (as well as importing backup)

Note that I used `os.path.dirname(os.path.abspath(__file__))` for filepaths in
backup. This is to allow the script to be ran even if the user isnt in the same directory
as the script. This is not implemented elsewhere. I will create an Issue and fix this later.